moving helpers to auth, add new users to db on first login
