+++ microservices/authentication/router/router.go
@@ -10,7 +10,6 @@ func InitAuthRouter(r chi.Router, auth *handlers.Auth, conf *jwt.GitCollabJwtCon
 	// add all routes and handlers below
 
 	r.Route("/auth", func(r chi.Router) {
-		//
 		r.Get("/redirect-url", auth.GithubRedirectHandler)
 		r.Post("/signin", auth.LoginHandler)
 		r.Get("/logout", auth.LogoutHandler)
